requirements.txt
pandas==2.0.3
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
